+++ src/changelog-updater.ts
@@ -6,7 +6,7 @@ import {DependabotEntry} from './entry-extractor'
 interface ParsedResult {
   foundDuplicateEntry: boolean
   foundEntryToUpdate: boolean
-  changelogLineNumber: number
+  lineToUpdate: number
   versionFound: boolean
   dependencySectionFound: boolean
   contents: string[]
@@ -107,10 +107,17 @@ function addNewEntry(
   // We build the entry string "backwards" so that we can only do one write, and base it on if the correct
   // sections exist
   let changelogEntry = buildEntryLine(prefix, entry)
-  const lineNumber = result.changelogLineNumber
+  const lineNumber = result.lineToUpdate
   if (!result.dependencySectionFound) {
     changelogEntry = `### Dependencies${EOL}${changelogEntry}`
+
+    // Check if the line number is last.
+    // If not, add a blank line between the last section and the next version
+    if (lineNumber < result.contents.length - 1) {
+      changelogEntry = `${changelogEntry}${EOL}`
+    }
   }
+
   writeEntry(lineNumber, changelogPath, changelogEntry, result.contents)
 }
 
@@ -119,7 +126,7 @@ function updateEntry(
   changelogPath: fs.PathLike,
   result: ParsedResult
 ): void {
-  const lineNumber = result.changelogLineNumber
+  const lineNumber = result.lineToUpdate
   const existingLine = result.contents[lineNumber]
   const existingPackage = existingLine.split(' to ')[0]
   const existingPullRequests = extractAssociatedPullRequests(existingLine)
@@ -190,7 +197,7 @@ async function parseChangelogForEntry(
   })
 
   let lineNumber = 0
-  let changelogLineNumber = 0
+  let lineToUpdate = 0
   let versionFound = false
   let dependencySectionFound = false
   let foundLastEntry = false
@@ -217,29 +224,43 @@ async function parseChangelogForEntry(
     }
 
     if (versionFound) {
-      // Inside version section
-      if (line.startsWith('### ')) {
+      // If we are finding a new version after having found the right version
+      if (line.startsWith('## ')) {
+        // Then we have found the last entry regardless of if we found the dependency section
+        foundLastEntry = true
+
+        // If we haven't found the dependency section, we need to set the line to update number
+        if (!dependencySectionFound) {
+          lineToUpdate = lineNumber
+        }
+      } else if (line.startsWith('### ')) {
+        // If we are finding a new section and we have found the right version
+        // and we have found the dependency section, we are moving into a new section
+        // Otherwise, see if this is the dependency section
         if (dependencySectionFound) {
           foundLastEntry = true
         } else {
           dependencySectionFound = DEPENDENCY_SECTION_REGEX.test(line)
-          changelogLineNumber = lineNumber + 1
+          lineToUpdate = lineNumber + 1
         }
       } else if (SECTION_ENTRY_REGEX.test(line)) {
         if (line.startsWith(entryLine)) {
+          // If we are finding a duplicate line, we have found duplicate entry and we will skip
           foundDuplicateEntry = true
         } else if (entryLineStartRegex.test(line)) {
+          // If we are finding the start to the entry, we have an entry to update and we will overwrite it
           foundEntryToUpdate = true
-          changelogLineNumber = lineNumber
+          lineToUpdate = lineNumber
         } else {
-          changelogLineNumber = lineNumber + 1
+          // Assume we find the last line if we find an entry
+          // We will append our new entry to end on the next line
+          lineToUpdate = lineNumber + 1
         }
-      } else {
-        foundLastEntry = true
       }
     } else if (versionRegex.test(line)) {
+      // If we have not found the version, see if this is the version
       versionFound = true
-      changelogLineNumber = lineNumber + 1
+      lineToUpdate = lineNumber + 1
     }
 
     lineNumber++
@@ -249,8 +270,8 @@ async function parseChangelogForEntry(
 
   // If we are at the end of the file, and we never found the last entry of the dependencies,
   // it is because the last entry was the last line of the file
-  changelogLineNumber = lastLineCheck(
-    changelogLineNumber,
+  lineToUpdate = lastLineCheck(
+    lineToUpdate,
     lineNumber,
     foundLastEntry || foundDuplicateEntry || foundEntryToUpdate,
     versionFound
@@ -259,7 +280,7 @@ async function parseChangelogForEntry(
   return {
     foundDuplicateEntry,
     foundEntryToUpdate,
-    changelogLineNumber,
+    lineToUpdate,
     versionFound,
     dependencySectionFound,
     contents
@@ -267,7 +288,7 @@ async function parseChangelogForEntry(
 }
 
 function lastLineCheck(
-  changelogLineNumber: number,
+  lineToUpdate: number,
   fileLength: number,
   foundLastEntry: boolean,
   versionFound: boolean
@@ -275,5 +296,5 @@ function lastLineCheck(
   if (!foundLastEntry && versionFound) {
     return fileLength
   }
-  return changelogLineNumber
+  return lineToUpdate
 }